Introduce state-based aggregation helper for partitioned tables
