+++ custom/icds_reports/utils/aggregation_helpers/distributed/ls_awc_visit_form.py
@@ -1,22 +1,17 @@
 from dateutil.relativedelta import relativedelta
+
 from custom.icds_reports.const import AGG_LS_AWC_VISIT_TABLE
 from custom.icds_reports.utils.aggregation_helpers import month_formatter
-from custom.icds_reports.utils.aggregation_helpers.distributed.base import BaseICDSAggregationDistributedHelper
+from custom.icds_reports.utils.aggregation_helpers.distributed.base import (
+    StateBasedAggregationPartitionedHelper,
+)
 
 
-class LSAwcMgtFormAggDistributedHelper(BaseICDSAggregationDistributedHelper):
+class LSAwcMgtFormAggDistributedHelper(StateBasedAggregationPartitionedHelper):
     helper_key = 'ls-awc-mgt-form'
     ucr_data_source_id = 'static-awc_mgt_forms'
     aggregate_parent_table = AGG_LS_AWC_VISIT_TABLE
     aggregate_child_table_prefix = 'icds_db_ls_awc_mgt_form_'
-
-    def aggregate(self, cursor):
-        drop_query = self.drop_table_query()
-        curr_month_query, curr_month_params = self.create_table_query()
-        agg_query, agg_param = self.aggregate_query()
-        cursor.execute(drop_query)
-        cursor.execute(curr_month_query, curr_month_params)
-        cursor.execute(agg_query, agg_param)
 
     def aggregate_query(self):
         tablename = self.generate_child_tablename(self.month)
