+++ custom/icds_reports/utils/aggregation_helpers/distributed/ls_vhnd_form.py
@@ -1,22 +1,17 @@
 from dateutil.relativedelta import relativedelta
+
 from custom.icds_reports.const import AGG_LS_VHND_TABLE
 from custom.icds_reports.utils.aggregation_helpers import month_formatter
-from custom.icds_reports.utils.aggregation_helpers.distributed.base import BaseICDSAggregationDistributedHelper
+from custom.icds_reports.utils.aggregation_helpers.distributed.base import (
+    StateBasedAggregationPartitionedHelper,
+)
 
 
-class LSVhndFormAggDistributedHelper(BaseICDSAggregationDistributedHelper):
+class LSVhndFormAggDistributedHelper(StateBasedAggregationPartitionedHelper):
     helper_key = 'ls-vhnd-form'
     ucr_data_source_id = 'static-ls_vhnd_form'
     aggregate_parent_table = AGG_LS_VHND_TABLE
     aggregate_child_table_prefix = 'icds_db_ls_vhnd_form_'
-
-    def aggregate(self, cursor):
-        drop_query = self.drop_table_query()
-        curr_month_query, curr_month_params = self.create_table_query()
-        agg_query, agg_param = self.aggregate_query()
-        cursor.execute(drop_query)
-        cursor.execute(curr_month_query, curr_month_params)
-        cursor.execute(agg_query, agg_param)
 
     def aggregate_query(self):
         month = self.month.replace(day=1)
